+++ CHANGES.rst
@@ -5,6 +5,12 @@ Change Log
 ==========
 
 
+UNRELEASED
+==========
+
+* logging: handler allows custom formatting of the *srcinfo* part
+
+
 v0.1.1 (2022-10-29)
 ===================